backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
requests
passlib[bcrypt]
python-jose
pydantic
pytest
httpx
pydantic[email]